requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas
pandas-stubs
configparser
prettytable
python-telegram-bot
azure-data-tables
azure-identity
azure-monitor-query
requests
aiohttp
python-dotenv
python-binance
pandas
python-kucoin
pyodbc
pycoingecko
twikit
feedparser
beautifulsoup4
pytest
pytest-cov
ruff
cryptography==43.0.3
plotly
google-generativeai
pypandoc